src/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.40.0
google-generativeai>=0.7.0
PyYAML>=6.0.1
